+++ routes/web.php
@@ -28,7 +28,11 @@ Route::group(['middleware' => ['auth', '2fa'], 'guard' => 'auth'], function () {
     Route::resource('users', 'UserController');
     Route::get('about', 'AboutController@index');
     Route::get('authlog', 'UserController@authlog');
-    Route::get('dependency-map', 'MapController@dependencyMap');
+
+    // Maps
+    Route::group(['prefix' => 'maps', 'namespace' => 'Maps'], function () {
+        Route::get('dependency', 'DependencyController@dependencyMap');
+    });
 
     // admin pages
     Route::group(['guard' => 'admin'], function () {
